Add support for reconnecting from broken pipe
When the connection goes down, for instance after suspend + wake up, or
due to network loss from mobile tethering, it is desirable that the
connection is re-established without the user having to explicitly
initiate a reconnection attempt.
To discriminate between above situation and when the server has been
shut down (in which case it seems senseless to try to reconnect),
a new connection status called RECONNECTING is useful.
When in RECONNECTING mode, we want to make one reconnection attempt
every n seconds, until we successfully have connected to the server.
Also, we want to remain in RECONNECTING mode until that happens.