Add support for buffer state persistence
Between sessions or from having been disconnected and then reconnected,
it is desirable to have the last displayed weechat-buffer displayed
again. The file state.py has a class State meant to keep track of what
buffer was displayed the last time. It also supports keeping track of
which server nodes in the treeview are expanded, so that persistence of
those as well can be implemented in the future.
When the application is closed, we want the state information to be
written to a file, so that we can read from that the next time we start
the application. This is done using the built-in pickle library.