Add expanded servers persistency
It is desirable that the servers in the buffer list treeview retain
their state of being expanded or collapsed between sessions, or after a
reconnect event happened.
In order to do that, one must save those states when i) the toplevel
window is closed, ii) the application is closed using the menu option
"quit" or iii) the network connection has failed.