Add support for synching the server hotlist
WeeChat's way of keeping track of buffer activity involves what
they call a hotlist, in which a priority value is attached to
each buffer. By requesting this hotlist when we connect to the
server, we can mark our buffers with the same notification-
levels that WeeChat itself shows. Also, we can tell WeeChat to
mark buffers as read, when we switch away from them. Last, we
can probe for the hotlist at a regular interval so that if
buffers are marked as read by WeeChat or another client, those
changes are propagated to our client as well.