Make possible to switch to dark theme on the fly
A checkbutton in the menu is added so that users can toggle using
dark theme variant, if it exists. If darkmode is toggled, the
color value for the foreground color is re-read from the css info,
and the notifyLevelChanged signal is emitted to trigger the color
values to be copied into the bufferlist model.