Use a fallback css when dark theme variant is missing
On Ubuntu 18.04, the default theme is Ambience, which lacks a dark
variant. Because of this, dark mode is not working. This commit solves
the problem by checking if the theme name is not Adwaita (one might instead
use a list of theme names that support dark variants), and if that is the
case, applies a fallback css style provider for darkmode.