Do not call show_all() when activated
In order to be able to use the hide() method for e.g. nicklist
widgets, it is necessary to avoid calling show_all() on toplevel
widgets, such as the main window, since show_all() reverts the
hidden status of all child widgets.
Instead of doing show_all() on the toplevel window each time the
application is activated, it is enough to do it once upon its
creation. We can then be sure that when calling hide() on a
widget, it will stay hidden unless we explicitly decide to
show it.