make Buffer a subclass of BufferWidget
Instead of the buffer widget being a member of the Buffer, the
Buffer will now inherit the BufferWidget. This makes it so that
the widget and the buffer can be seen as one and the same. The
reason is to decrease some of the cognitive load associated
to keep track of all different subcomponents related to a
buffer. The code should also be easier to read.

A new method get_url_tag is added because the parent class,
BufferWidget needs access to the url TextTag, only available
from the child class which creates the TextBuffer. The child class
thus must implement the method in order for the parent class
to function.